[SH] Bug Fix: Can now correctly boot 32-bit kernels on the STb7109.
Fixed a bug with the PMB setup, whereby U-boot was leaving PMB entries
valid, that clashed with those used by the linux kernel. The fix is just
to invalidate all the PMB entires, modulo #0, which is used to map the
kernel itself. i.e. ensure PMB[n].V = 0, for n==1..15 prior to booting
the kernel.

This problem was only observed on a MB442+7109c3, but could affect
other platforms.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>