[SH] Initial (snapshot) support for USB mass-storage on MB442.
This is the initial snapshot version of USB mass-storage for the SH
architecture. This has only been tested on one ST board, the MB442.

It should be noted that this version disables the data caches when the
code is activly talking to the USB bus. There are also several outstanding
tasks, which are labeled with a 'QQQ' comment.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>