[Experimental] Performance improvment for OHCI USB.
Uncommenting the following line will remove some delays in the USB
sub-system. However, it is possible that some of these removed delays
may be actually necessary on some hardware devices.
So, although this gives a useful performance improvement, it does
so at the potential risk of reliability and stability.  This should be
considered as an experiential configuration, and not at all recommended
for production or deployment code, unless it has been thoroughly
tested for your specific hardware.
Enabling this optimisation makes some assumptions, that may be totally
without any foundation.  If reliability or stability becomes an issue,
then please disable this configuration option, as a first step.
Use it at your own peril - caveat emptor!
//#define CONFIG_USB_AGGRESSIVE_MINIMAL_DELAYS

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>