[SH] USB now works on both STb710x and STi7200 chips.
Added support for USB Mass-Storage for the STi7200 chip. It should be
noted that although the STi7200 supports 3 USB Host Controllers, that
only one may be used. This is a current restriction of the generic U-boot
sources, and not the SH port.

Also now tested on several different boards, including the following ST
boards: mb519, mb442, mb448, hms1.

Also tidied up the initialisation of USB (and SATA).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>