Fix to CFI flash, to support multi-bank devices better.
Fixed a problem that occurs with multi-bank devices, whereby, a sector was
in the wrong mode, and the wrong data was accessed. Now an explicit call
is made to ensure the sector is in the correct mode prior to accessing it.

Witihout this fix, the "protection" status information for a given sector
was indeterminate.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>