[OHCI] Fix a putative DataToggle Error whilst performing BULK transfers.
There was a subtle bug with the way that the OHCI data toggle carry
bitfields were managed. It was possible for the HC and the remote device
to get out of step with the data toggle values, due to the fact that
the ED structures were re-used, without saving and restoring the ED.C
bitfield across this re-use.

This problem was only observed when there were multiple (strictly greater
than one) BOT USB Mass-Storage devices present on the system. The system
would sometimes hang, and report timeouts. Simply adding or removing
a device, or exchanging the positions on the bus of two devices could
invert the behaviour. It initially looked like a random instability.

Due to the fact that only a statically defined small number (eight) ED
structures were available to the HCD to manage all the devices, the EDs
were regularly re-used, if multiple devices were present. This resulted
in the ED.C (toggleCarry) bitfield potentially losing its value between
2 consecutive BULK accesses to a single device, if the ED structure was
temporarily re-used to communicate with a different device. The following
simple command sequence was enough to provoke it: "usb reset ; usb part"

This patch saves the ED.C bitfield in the usb_device structure, after
each transaction. It will then invert the hardware visible ED.C for new
BULK transfers, if a DataToggle error would have otherwise resulted.

It should be noted that CONTROL transfers were immune from this problem.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>