[SH] Added initial support for NAND FLASH on the ST40 Silicon.
Added initial support added for NAND FLASH for the CB101 board with
the STx7200 silicon from STMicroelectronics. This has been tested for
both 29-bit and 32-bit environments, but only on the CB101, with both
(on-board) NAND chips. Testing on the MB519, and other boards to be done.

Testing included using the JFFS2 filesystem, and the "mtdparts"
environment variable to define partitions for both NAND and NOR
partitions.

Performance of this initial version is sub-optimal. This version takes
no advantage of the caches that are available on the SuperH architecture.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>