[SH] Added support for the MB618 (STx7111) MBoard
This update is to add support for STMicroelectronics MB618 MBoard for
the STx7111 chip.

This update excludes support for one significant block - the GMAC.
The GMAC (GigaBit Ethernet MAC) requires a modified driver from the
previous generation 10/100Mbps ST-MAC blocks, as the block's register
set has changed. Currently, this modified driver does not yet exist.

With the exception of the GMAC, this update is believed to be complete,
although further testing is required. This patch includes support for
serial, USB, NAND & NOR flash.  It has been tested on both 29-bit and
32-bit (SE) modes; both from GDB and booting from NOR-flash.

It should be noted that the MB618 supports both NOR and NAND
flashes. However, only one may be visible any any one point. Hence,
it is necessary to re-configure and re-build U-boot, to swap between
NAND/NOR flash devices.

Several minor changes to other files (mostly configuration) have been
made.  These were just to keep the code as "unified" as possible.

Finally, this is the first update that includes support for the SH4-300
series family of CPU cores.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>