[SH] Ensure MMU coherency (with RTE) on SE mode-switching
Previously the code executed 8 NOP instructions, immediately after
entering 32-bit mode (SE=1).  However, this is NOT architecturally
guaranteed (on SH4-300 series cores). Executing an RTE instruction
instead, is guaranteed to preserve the MMU coherency, on all SH cores.

This patch, replaces the eight NOPs, with an RTE instruction, to ensure
the MMU coherency, when mode-switching.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>