[SH] Use the on-chip Watchdog to effect a (power-on) reset
Prior to this change, the "reset" command was implemented by setting
SR.BL=1, and issuing a TRAPA instruction. This works for for all SH
devices, except for the SH4-200 series cores in 32-bit (SE) mode.

This update is to use the on-chip watchdog timer instead, to effect a
power-on-reset. This modification will also be applied to devices in
which the TRAPA idiom would have worked. That is, all "reset" commands
for the SH family, will now use the watchdog as the reset mechanism.

As a result of this change, the chip-specific st?7???_reset() functions
have been all been replaced by a single sh_reset() function, which now
resides in lib_sh/board.c.

Finally, due to a previously unknown problem with the MB618, which meant
that warm boots (from FLASH) hung, the "poke table" has been modified
to comment out some pokes which appear to be benign during a cold boot,
but result in a hang during a warm boot.

This code modification has only been tested on the MB618 & MB519 boards.
Further testing is required.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>