[SH] Completed support for the MB618 (STx7111) MBoard
This update includes support for the new GMAC Gigabit MAC used in
STMicroelectronics products, and found in the STx7111 device.

This has been tested on both 29-bit and 32-bit (SE mode) environments.
Both "ping" and "nfs" seem to work with no problems. As I do not have
access to a Gigabit PHY, this has only been tested at 100Mpbs.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>