[SH] Added support for all 256MB of LMI0 RAM on MB671 in SE-mode.
The MB671 board has 256MB of memory in LMI0. Prior to this update, U-boot
only supported the first 128MB of this memory in SE (32-bit) mode. This
meant that both the kernel and initrd ramdisks needed to be located in
the first 128MB of LMI0.  Also, it meant that the "mtest" command only
tested just under half of the LMI0 memory.  This update has refactored
the PMB code, such that there are now two contiguous 128MB PMB entries
(instead of just the one) to map the entire 256MB of LMI0 memory. As a
result, the aforementioned restrictions are now removed. As a result,
U-boot will now be located at 0x8FF00000 on the MB671.

The PMB refactoring has resulted in the following changes:

* PMB[0:1] is used to map main memory, and will mostly be used as cached,
  except on initialization, and when passing control to the linux kernel,
  where it will be un-cached.

* PMB[2:3] is an alias for PMB[0:1], except it is guaranteed to always
  be UN-cached.

* PMB[0:1]  will map VA 0x80000000 .. 0x8fffffff.  (256MB)
  PMB[2:3]  will map VA 0x90000000 .. 0x9fffffff.  (256MB)
  PMB[4:15] will map VA 0xa0000000 .. 0xbfffffff.  (512MB)

* If sizeof(LMI) <= 128MB, then PMB[1] and PMB[3] are unused,
  i.e. PMB[1].V == PMB[3].V == 0.

  see include/asm-sh/pmb.h for more details.

In addition, a new command "pmb" has been added, which when run, will
display the state of all the currently valid PMB entries, in SE-mode.

Also, a miscellaneous tidy up of some mis-spellings.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>