[SH] Added support for the CB102 (STx7200) Board.
This update includes proper support for the CB102 board, which
includes STMicroelectronics' STx7200 chip.

This version has been tested on both 29-bit and 32-bit modes. This
includes support for accessing all 256MB of LMI0 RAM in 32-bit mode.

Successful testing has been performed as follows:

	ASC (UART#3, ttyAS0)
	NOR-flash (8MB, M29W640GB)
	Ethernet (STMAC + LAN8700)
	USB (front & rear) [OHCI only]
	Booting both 29-bit and 32-bit linux kernels from NOR.

Also some additional descriptive symbols are defined in cpu/sh/start.S,
which should facilitate debugging of this file.

Fixed a minor bug in lib_sh/sh_linux.c, whereby the wrong mode was
reported for both linux & U-boot, if the kernel was inappropriate.

Finally, the stm-stmac.c driver has been updated to be tolerant of
revision changes in the PHYID register for the LAN8700 and STe101p.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>