[generic] Pad printed memory sizes with a field width of 3.
The function print_size() will now pad memory sizes to a field width of 3
(it was 2). This is required as a lot of boards now have > 99MB of RAM
on them, and the vertical alignments were inconsistent, when displaying
the capacity for a board with a lot of RAM, but a small amount of flash.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>