[SH] Completed support for the MB671 (STx7200) MBoard.
This update includes the full release for the MB671 Mboard, which uses
STMicroelectronics' STx7200 chip.

This update builds on the previous updates for this board, and in
addition has now been properly tested with both of the on-chip ST-MAC
blocks. Thus, can can use either the on-board SMSC PHY, or an off-board
PHY, such as the one on the MB539B daughter board. It should be noted
that the on-board PHY (on MII0) is in RMII mode, whereas, the off-board
PHY (on MII1) is in MII mode.

As this board does not contain any on-board NAND flash, there is *no*
NAND flash support provided with this board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>