[ASC] Bug-fix and refactoring of the (STM) ASC serial driver
This update is a refactoring of the drivers/serial/stm-asc.c ASC driver
file for the UART used by STMicroelectronics. This was primarily to
address some confusion/misdirection over the accuracy of the comments,
which had been reported a few times.  As a result, the comments are
now correct, and there is a comment for each function, which should
significantly address the maintainability of the file.

As a result of the tidying up, some functions have been deleted, and
(in some cases) recreated in a different guise, further improving the
readability and maintainability of the driver.

Code has been commented out, removing the support for the anachronistically
slower baud rates (below 9600). If required, it can be re-enabled!

Finally, a bug was fixed which had previously resulted in baud rates
slower than 19200 being mis-configured with the wrong baud rate divisor.
This has now been fixed, and all supported baud rates have been tested
(on a mb628 board).

The baud rates now supported are: 9600, 19200, 38400, 57600, 115200.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>