[SH] Phase I support for STMicroelectronics' MB628 (STx7141)
This update provides initial (Phase I) support for the MB628 board,
which uses STMicroelectronics' STx7141 SoC device. This version is not
intended to be generally useful, as it is engineered to be targeted at
a very specific requirement, to facilitate an alternative debug solution
for the STx7141 platform. As such is has several omissions and caveats.

This update, only provides support for the MB628 in 29-bit (not 32-bit)
mode; for the eSTB core (not eCM); and will only support RAM, NOR
Flash and serial (ASC) devices. That is, there is no support for USB,
Ethernet, NAND flash, SATA, PCI, or any other IP block. In addition,
this version does *not* perform any of the traditional GDB "pokes" to set
up LMI/EMI/Clocks etc. As such, it is expected that the other ST40 core
(eCM) will perform all the LMI/EMI/Clock setup, and then start the eSTB
core at the address of U-boot in NOR flash.

This has been given minimal testing, but tested as above on one board,
and successfully booted Linux from NOR flash.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>