[SH] bug fix to include/asm-sh/st40reg.h (ST40_EMI_BANK_ENABLE)
Fixed a bug that the register offset for ST40_EMI_BANK_ENABLE was
incorrect.  Instead, an unused macro ST40_EMI_BANKNUMBER had this offset,
and this (unused macro) has now been removed.

This bug was originally reported in STMicroelectronics' bare-machine
tool-set as DDTS #INSbl28111.

This bug fix is required with boards that need to disable some of the
EMI banks, due to a lot of NOR flash being present on the board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>