[MII] Print out the OUI with a field width of 6 (was 4)
The OUI is a 24-bit value, and so it requires a maximum of 6 character
positions to print it in hex. Previously it was only 4 character
positions. PHYs which have any bits set in the top octet resulted in
the output of the "mii info" command being unaligned. This update fixes
that annoyance.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>