[SH] Include support for latest EPLD revisions on MB618 (STx7111) boards.
The STMicroelectronics' MB618 (STx7111) board has undergone several EPLD
revisions in the firmware, and the previous version of U-Boot became
incompatible with the newer versions.

This update is to address that, by supporting version 06 (and later)
of the EPLD firmware. As a result of this change, all versions of the
firmware prior to version 06 (specifically 04 and 05) are no longer
supported.  Boards with earlier versions of the supported firmware must
upgrade to a later version to use this patch to U-boot.

The new EPLD firmware has changed the memory map of the EPLD itself
from physical address 0x04000000 to 0x06000000. In addition, the test
and control registers are now "banked".  Also, we now print both the
EPLD version number, and the board revision number, during initialization.

Finally, a new configuration macro "CONFIG_SH_NO_EPLD" now exists, which
can be used to instruct U-Boot to disregard the presence of a EPLD.
This is useful for clones of the MB618, which do *not* have a EPLD.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>