[SH] Bug-fix to $monitor_sec for the STMicroelectronics' MB618 board.
Bug-fix to include/configs/mb618.h.  This fix will now set the correct
sector specification for the $monitor_sec environment variable, so that
the "run unprot" and the "run update" scripts will run correctly. Prior
to this fix, the code assumed that the default fitted NOR flash geometry
was "uniform", whereas it actually has a "bottom" configuration.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>