[SH] Complete support for latest EPLD revisions on MB618 (STx7111) boards.
This update is a bug-fix to correct an omission that was made during a
recent commit, whose one line summary was:

    "[SH] Include support for latest EPLD revisions on MB618 (STx7111) boards."

An omission was made in that update, as one source file that was
logically part of that commit was inadvertently omitted from the
referenced commit set.  The missing file was:

	board/mb618/init-mb618.S

This update just adds that missing file.  Without this update, then
U-Boot would not boot from flash correctly.  This update should be
treated logically as if it were part of the related commit.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>