[Fix] Stop corruption due to incorrect configuration of CFG_MAX_FLASH_SECT
The configuration macro "CFG_MAX_FLASH_SECT" should be set to the largest
number of sectors in any NOR flash device on the system. However, the
program did not check if this assumption was violated or not.  If this
assumption was violated, then this results in corruption of memory
(overwriting data), and (at least on my board) a "hang".

This update is a simple test to ensure that the number of sectors probed
is not more than CFG_MAX_FLASH_SECT, and avoids the potential array
bounds violations the otherwise would occur. The array bound violations
pertain to the following 2 fields:

	flash_info_t.start[CFG_MAX_FLASH_SECT];
	flash_info_t.protect[CFG_MAX_FLASH_SECT];

If the test catches such a error condition, then it prints a suitable
diagnostic, and sets the FLASH size to be zero bytes.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>