[SH] First pass of initial support for the STi7105-SDK Board.
This update includes initial support for STMicroelectronics' STi7105-SDK
board, which uses STMicroelectronics' STx7105 SoC device.

This includes support for the SoC on this board, as well as the following
blocks: Serial (ASC0), LMI DDR, NOR Flash, USB (OHCI).  The following
are not supported: ST-MAC, NAND Flash, Serial Flash (SPI), PCI, SATA,
EHCI, or any of the other blocks that STMicroelectronics does not support
in U-boot.

All development and testing has been performed with the STi7105-SDK-MB
mated with a STi7105-SDK-SB board. Changes to the configuration
file (include/configs/pdk7105.h) and possibly other changes will be
necessary, if the STi7105-SDK-MB is not mated with a STi7105-SDK-SB board
(specifically, serial). The stand alone configuration (STi7105-SDK-MB
only) has not been tested.

This update has been tested in both 29-bit, and 32-bit (SE) modes.

NOTE: The file "init-pdk7105.S" is a copy of the MB680 version, and hence
needs to be updated with the latest version of the target packs, which
(at the time of writing) are still undergoing revisions) for the correct
board. As a result, this patch has not been tested in a "boot-from-NOR"
environment.  Updating the "init-pdk7105.S" is fully expected to yield
a good bootable system.

Other limitations:

It should be noted that although the board has 512MiB of DDR on the LMI,
due to the static nature of the PMB configuration in U-boot, that only
the first half of the LMI is mapped by the PMBs in U-Boot. This is due
to the fact that the entire PMB region is only 1GiB in size, and U-Boot
(currently)requires to have the LMI memory mapped *both* cached and
un-cached. Although, this is possible, it excludes being able to map FLASH
in the PMB as well, which is unacceptable.  Hence the compromise is that
U-boot can only see the first 256MiB of RAM. This is not a real problem,
as when the kernel starts, it will be able to see 100% of the RAM. So,
the only real restriction is whilst running U-boot, users must load the
kernel + ramdisk into the first 256MiB, which is not really a problem!

Only the first 96MB of NOR flash is visible. This is due to the fact that
the *entire* EMI space is only 128MiB in size, and a 128MiB NOR flash is
instantiated. Again, it would be logically possible to map all of the NOR
flash, but that would mean that no other EMI devices could be accessed,
such as: NAND Flash, CI, PCI. So, this compromise seems reasonable.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>