[SH] Second pass of initial support for the STi7105-SDK Board.
This update includes further support for STMicroelectronics' STi7105-SDK
board, which uses STMicroelectronics' STx7105 SoC device.

Specifically, this update now includes support for "boot-from-NOR"
environment. That is, the file "init-pdk7105.S" has now been updated to
include an automatically generated "romgen-pdk7105.inc" file using the
ROMGEN utility. This has been tested, and shown to perform a boot from
NOR flash on both 29-bit and 32-bit modes.

It should be noted that the "reset" command does not appear to result
in a proper boot, so further work is required to investigate this.

The other restrictions mentioned in the previous update still apply.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>