[SH] Added 32-bit mode (SE) support for the MB628 (STx7141)
This update adds 32-bit (Space-Enhanced) support for STMicroelectronics'
MB628 Mboard, which uses STMicroelectronics' STx7141 SoC device.

This has only been given minimal testing, and largely still has the same
set of caveats and restrictions as the previous update for the MB628.
Specifically, this update does not include the set of GDB pokes to make
this self-bootable, and so still relies on another system to perform
the necessary pokes.

The memory map has been revised, reflecting the latest version. Hence
the eSTB now has 50MiB starting 8MiB from the beginning of the LMI.
The current memory map is now in board/mb628/config.mk, and follows:

	Region		29-bit		32-bit		Size
	------		------		------		----
	LMI_BASE	0x0c000000	0x40000000	128MiB
	VIDEO_BASE	0x0c000000	0x40000000	  4MiB
	AUDIO_BASE	0x0c400000	0x40400000	  4MiB
	eSTB_BASE	0x0c800000	0x40800000	 50MiB
	eSTB_END	0x0f9fffff	0x439fffff	   ---

All the above addresses are *Physical* Addresses.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>