[SH] Add support for the serial console via DTF over a JTAG link.
Adds support for emulating a serial console using STMicroelectronics'
DTF (Data Transfer Format) mechanism, over a JTAG link, to a remote GDB
debugger on a host machine. This may be used instead of a physical UART.

As U-boot expects to drive real physical UART, there are some implications
of using DTF instead of a UART. The known issues are as follows:

There is an unwanted echo of all read characters, thus it is more
visually difficult to see what command U-Boot is about to process.

Some of the keys are interpreted by GDB first, so for example, "Control-C"
is not passed to U-boot, but the GDB command interface takes control of
the keyboard again, suspending U-boot.  The "Delete" and "Backspace" keys
do not work (although Control-H does), resulting in it being difficult
to edit commands if they are mis-typed.  Similarly, the "Arrow" and
other keys with escape sequences are not displayed correctly, making it
difficult to see the edited command-line that U-boot sees.

The serial I/O console interface over DTF+JTAG (as implemented) is fairly
slow (noticeably slower that a real serial console). It should be possible
to speed it up if required, but this will require further effort.

Using this DTF+JTAG interface requires GDB to be running and connected
to the host machine. If there is no GDB connected, then U-boot will hang
waiting for GDB to "talk to it" for each and every character read/written.

To enable serial console over JTAG, then modify your config.h file to:

Alternatively, the following allows easy toggling between using DTF
and the physical ASC UART:

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>