[SH] Added "ROMGEN" files for the CB103 Board.
This update only adds the "ROMGEN" files for the CB103 board.  It should
be noted that, this is *not* a port for the CB103, as is is *only* the
romgen files that have been added. Until there is a proper port for the
CB103, then it is expected that these files are copied into the tree
for the CB102, and a CB102 it built using the CB103 pokes.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>