[SH] Second cut of Boot-from-NAND for STx7105 (MB680)
This is the second GIT commit providing Boot-from-NAND flash for
STMicroelectronics' STx7105 device on the MB680B Mboard. This work is
still incomplete, and should be considered as a "pre-alpha" update.
There are some *significant* caveats and restrictions with this version.

This has only been given minimal testing, on a single MB680B board,
with cut 2.x silicon.  Testing, so far, has only been in 29-bit mode.

The current status, is that U-boot can now burn into NAND flash a binary
image, which complies with the ECC HW for Boot-Mode, that is 3 bytes
of ECC data for each 128 byte record. The resultant image is capable
of booting from a NAND flash device. This is a great point to get to,
but there is a lot more work to be done ...

Currently, it is possible to burn "u-boot.bin" to NAND flash, and boot
U-boot on power on. It should then be possible to load a vmlinux +
ramdisk from NAND to RAM, and boot linux.

Outstanding Issues:
------------------

Currently, in U-boot, only "bit-banging" is supported. This is sufficient
to read, write and erase NAND flash. However, once booted, and the NAND
controller IP is in "boot-mode", then "bit-banging" no longer works -
looks like conflicts with the NAND controller IP block over ownership
of the chip select line. This is now a known H/W bug in the NAND
controller. As a result, there are 2 key issues:

	a) A NAND booted U-boot can NOT read/write NAND with bit-banging.

	b) All reads/write from/to NAND must *not* be in NAND boot mode.
	   (i.e. boot-from-NOR mode).

To solve the above issues, U-boot ultimately needs to support "Flex Mode"
(or Advanced Flex Mode), removing the need to use any bit-banging at all.

There is no ECC correction yet. Any bit errors will currently be ignored.

It looks as if the image in NAND flash must not be re-mapped. That is
Logical Block == Physical Block. Analysis suggests that only logical
block zero is "remapped", and not the following blocks! This can be
worked around by burning "u-boot.bin" to physical block zero.

U-boot somehow needs to "partition" the NAND device into 3 regions:
		1) "boot-mode"-region
		2) "filesystem"-region
		3) Bad Block table region

	U-boot needs to know when to use the correct ECC algorithm for
	each region, and where the region boundaries are. Specifically,
	in boot-mode there are 3 ECC bytes per 128-byte record, whereas in
	the "filesystem"-region there are 3 ECC bytes per 512-byte page.

Only small-page (512 bytes) NAND devices have been tested.

Some of the code in board/mb680 needs to be moved from a board-specific
directory, to either a chip or cpu tree.

Need to test/support 32-bit mode.

Need to try on Rev C of MB680.

Only tested on the STx7105, need to work on other devices/boards.

Still a lot of tidying up, and removal of "QQQ" code/comments.

Finally, the MB705 board, to which the MB680 is mated, supports
boot-from-NAND by *swapping* the chip select lines (CSA and CSB).
There are some ramifications to this that have not been fully considered.
Specifically, the sizes of the 2 banks are not swapped by anything.
Thus, when booting from NAND, only 8MiB of NOR is visible. Either the
target-packs; the romgen generated pokes; or U-boot initialization will
probably need to adjust this!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>