[SH] Updated for STMicroelectronics' STx7105 cut 2.x on MB680
This update adds support for cut 2.x of the STx7105 SoC device, on
the MB680 MBoard. Specifically, this update uses the latest "romgen"
generated set of GDB "pokes", and modifies the code to take account of
recent silicon changes in cut 2.x pertaining to the configuration of
the ST-MAC ethernet.

It should be noted that set of GDB pokes for cut 1.x is also included
with this update, such that if users wish to, then they may use either
the cut 1.x or the cut 2.x pokes. Selection is made by editing a '#if'
directive in "board/mb680/init-mb680.S".

Note: this updated patch will (as is) break ethernet support on cut
1.x silicon.  This patch should only really be used on cut 2.x, unless
the file "cpu/sh/stx7105/stx7105.c" is regressed to the previous version.

Testing for this has been made on a single MB680B+MB705B board. Testing
included booting from NOR flash in both 29-bit and 32-bit modes.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>