[SH] Add compile-time diagnostic if missing PMB entries are found.
Add a #error diagnostic during compile time, if U-boot is configured in
32-bit mode to have more than 128MiB of main memory (CFG_SDRAM_SIZE),
but insufficient PMB entries are defined.

There is a common failure mode, whereby users will sometimes increase
CFG_SDRAM_SIZE, and not make the appropriate changes to "cpu/sh/start.S"
to add the necessary PMB definitions for SE-mode. So, this change is an
attempt at trying to catch this problem as early as possible.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>