[SH] Added "boot-from-NAND" capability for the PDK7105
This update properly merges the PDK77105 and MB680 "boot-from-NAND"
support, and provides the (restricted) capability on both boards.

The code was re-structured slightly, to migrate things from a
board-specific file to a chip/cpu specific file.

Note: this version has *not* been tested, other than to ensure the NAND
was written in "boot-from-NOR" mode. So, there may be a real problem with
this code, as far as real NAND booting is concerned.

This code has only been tried on the MB680 and the PDK7105.

There are still many outstanding issues with this version. Including
unusable BBT, wrong ECCs, no ECC correction...

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>