[SH] Speed up scanning for bad blocks in NAND on STMicroelectronics' boards.
Added a structure "stm_nand_badblock_pattern", and used it on the
current NAND-bootable boards (MB680+PDK7105) to only scan the first 2
pages of each block, and not every page in each block.  This provides a
significant performance improvement (x30+). This brings the U-boot NAND
scanning code in line with that used in linux. The speed up only applied
when U-Boot was asked to "create" a BBT (in the last 4 blocks).

This code has not been tested properly, and should not be considered
correct, although the (minimal) testing that has been done, suggests
this delta should be good.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>