[SH] Added "boot-from-NAND" for PDK7105 & other NAND-booting improvements.
This update is primarily aimed at the "boot-from-NAND" capability on
STMicroelectronics' ST40-based STx7105 SoC devices. In particular it is
targeted at the MB680 and the PDK7105 boards.

The main changes this patch provides are:

Support added for the PDK7105 (aka SDK705) board.

Support for both small and large page NAND devices is now complete
and correct.

There is now a "boundary" in the NAND device, below which we will
generate 3 bytes of ECC per 128 byte record, compatible with the
ST40 NAND controller H/W IP block. Above (and at) this boundary, the
traditional 3 bytes of ECC per 256 byte record is used. All ECC is
generated/corrected in software.  The boundary is defined by the macro
CFG_NAND_STM_BOOT_MODE_BOUNDARY.

The use of "Bad Block Tables" (BBTs) in boot-from-NAND is now mandatory
due to the requirement to have more than one ECC algorithm on a single
NAND device.  The previous update has the property that BBT generation
was awkward, and resulted in the need for a 3rd version of U-boot, just
to create it! This update fixes this issue, and BBT generation is now
simple and automatic, requiring no special treatment.

The previous version did not correct any ECC errors, this version will
now detect and correct all 1-bit errors.

On the MB680 (and the PDK7105), the Ready/Not-Busy signal is used instead
of a timed delay.  This should increase performance, and reliability.

Some code has been re-factored (i.e. moved about and renamed).

There were some bug-fixes and other improvements.

NOTE: This version still has some internal temporary development
debug code/diagnostics in it, which will be removed in the "official"
release. (Typically this development code is qualified by the string
"QQQ".)

There are now 3 macros in include/configs/<board>.h that need to be
modified to enable NAND booting:
	CFG_NAND_ECC_HW3_128
	CFG_BOOT_FROM_NAND
	CFG_NAND_STM_BOOT_MODE_BOUNDARY

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>