[SH] First cut with NAND FLEX-mode support for ST40 NAND Controller
This is the first commit with working support for the "FLEX-mode" NAND
controller IP on STMicroelectronics SoCs. This update only includes
support for the STx7105 device with cut 3.0 silicon (older cuts are
*not* supported).

This update includes support for the MB680+MB7105 and the PDK-7105 boards.
However, is has *not* been properly tested, and as such this is really
just a snapshot commit, prior to starting to do real testing. Testing
so far has been limited to a single MB680B+MB705B only in 29-bit mode.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>