[SH] Bug-Fix for the PDK7105 for boot-from-NAND
After testing the previous update, it was apparent that the base addresses
(and CSn numbers) were incorrect for the PDK7105 board. This update
corrects those errors, and now results in a properly working PDK7105.

Testing so far has been on the MB680B+MB705B & PDK7105-V1.2 boards.
Testing checked the ability to read and write to both NOR and NAND flash
in both 29-bit and 32-bit modes on both these boards. Also, testing
verified that both boot-from-NOR and boot-from-NAND both work! (on one
instance of each of these two boards)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>