[SH] Fixed the "reset" command, so it now resets the MB680 board.
Previously, the "reset" command on STx7105 SoC boards failed to reset
the board, and just hung, as the reset loop-back was "bypassed". This
update allows the reset chain to propagate.

This has been successfully tested on a MB680C board.

However, this does *not* appear to work on the PDK7105.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>