[SH] Added support for STx5197-CAB (including Boot-from-SPI)
This updates adds support for STMicroelectronics' STx5197 SoC device
(15x15) on the STx5197-CAB (cable) reference board. One of the key
additions in this new port is support for SPI serial flash - reading,
writing & booting.

This updated includes support for, and was tested using: 29-bit & 32-bit
(SE) modes; USB (OHCI); Ethernet; Serial console; M25P128 SPI serial
flash devices; boot from SPI.

It should be noted that if swapping between using the ST and Atmel Serial
Flash devices, that the configuration file (include/configs/5197cab.h)
needs to be updated, and U-boot re-built. Also, in order to boot from
the Atmel serial flash, that it needs to be in 512-byte page size - see
the aforementioned configuration file for details on how to effect this
one-time re-programming of the Atmel AT45DB321D serial flash on the board.

In order to read/write from/to the SPI serial flash, it should be noted
that the "eeprom read" and "eeprom write" commands should be used (not
the traditional NOR flash "cp", nor "md").

Currently, in this release, the U-boot environment variables are not
stored to the serial flash (using the "saveenv" command). For the
time being changes to the environment variables need to be made in the
aforementioned configuration file, and re-built & re-flashed.

All development and testing was done on the STi5197-CAB-V1.2A-068 board.

Finally, if a new serial flash device is to be substituted, then U-boot
may need to be "tweaked" to add support for it (cpu/sh/stx5197/spi.c).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>