[SH] Add "saveenv" support for SPI serial flash for the STx5197.
This updates adds support for U-boot's environment variables to be stored
in SPI serial flash on STMicroelectronics' STx5197 SoC devices on the
STx5197-CAB (cable) reference board, and the MB704 M-board.

The previous update for the STx5197 provided support to read and write
to SPI serial flashes, but it does not allow storing the environment
variables on that media. This update rectifies that omission.

This update has been tested in both 29-bit and 32-bit modes, on a 5197CAB
(V1.2A-068) board (although, not the MB704).

It should be noted that (at least on my 5197CAB board) that booting
requires the JTAG cable to be removed in order to boot successfully.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>