[SH] Preparation for boot-from-NAND with "skipping" Bad Blocks.
This update is just some ground-work for providing a "skipping" aware
U-boot. That is, preparing the code to allow U-boot to boot from NAND
with bad blocks near the beginning of the NAND, such that u-boot.bin is
burned with "holes", and the U-boot bootstrap re-locator needs to "skip"
over these bad blocks.

This update tidies up some minor issues with the linker script
files. Specifically:

	* There is now only a single "u-boot.lds" file for all SH SoCs,
	  (previously there was one per board).

	* A (seemingly benign) issue that "common" data was not
	  explicitly initialized, has now been fixed.

	* All the early initialization code (& data) that is needed
	  prior to re-locating U-boot into RAM has all been moved into
	  new ".text.init" (and ".data.init") sections. The u-boot.lds
	  file has been updated to "collect" these new "init" sections
	  to the beginning of the final ".text" section.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>