[SH] Add support for "skipping" bad blocks when booting from NAND.
If we want to store "u-boot.bin" in NAND flash starting at physical
block #0, but there are Bad Blocks in the first few blocks that we need
to "skip" over, then we can now configure U-boot to allow skipping of
these bad blocks, to allow u-boot to relocate itself correctly.

NOTE: Physical Block #0 *must* always be GOOD.

A new structure has now been added at offset 0x00C0 from the beginning
of physical block #0, which if present, will allow U-boot to "skip" over
bad blocks when re-locating itself from NAND to RAM. This structure has
the form:

	struct skip_bb
	{
		char    pattern[8];
		u32     block_size;
		u32     num_blocks;
		u32     array[num_blocks];
	};

	Where, "array[]" is a set of flags, whose value is interpreted as:
		if array[x] == 0,
			then physical block #x is GOOD
			else physical block #x is BAD
	and "pattern" holds the string "SKIP_BBs",
	and block_size is the NAND (erase) block size,
	and num_blocks is the number of entries in "array[]".
	see "cpu/sh/start.S" for details.

To help clarify, the following assertions should hold:

	assert( strcmp(pattern,"SKIP_BBs") == 0 );
	assert( (block_size % (16<<10)) == 0 );
	assert( sizeof(array) == num_blocks * 4 );
	assert( array[0] == 0 );

By default the linker will fill "array[]" with ZEROS (assume all blocks
are GOOD).  It is the responsibility of the configurer to ensure that
"array[]" is big enough.  It is the responsibility of the NAND flasher
to fill the array appropriately.

The following 3 macros are used to configure the "skipping" feature:
	CFG_NAND_SKIP_BAD_BLOCKS_ON_RELOCATING	/* to enable support */
	CFG_NAND_SKIP_BLOCK_SIZE		/* erase block size */
	CFG_NAND_SKIP_BLOCK_COUNT		/* size of "array[]" */

See "cpu/sh/start.S" and "include/configs/mb680.h" for details.

This update has been tested on a STx7105 (MB680C) using small-page
NAND, in both 29-bit and 32-bit modes. Testing was performed by burning
u-boot.bin with "holes" in it, and patching "array[]" to mark the holes
as if they were bad.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>